Rename: ui 를 view 로 이름 변경
